refactor editor context
