+++ frontend/dialob-composer-material/src/App.tsx
@@ -1,8 +1,9 @@
-import { ComposerProvider, EditorProvider } from './dialob'
-import multiPageForm from './dialob/test/multiPageForm.json';
-import ComposerLayoutView from './views/layout/ComposerLayoutView';
+import { ComposerProvider } from './dialob'
+import { EditorProvider } from './editor';
 import { IntlProvider } from 'react-intl';
 import messages from './intl';
+import multiPageForm from './dialob/test/multiPageForm.json';
+import ComposerLayoutView from './views/layout/ComposerLayoutView';
 
 function App() {
   return (
