+++ frontend/dialob-composer-material/src/dialob/actions.ts
@@ -1,9 +1,4 @@
-import { DialobItemTemplate, ValueSetEntry, ContextVariableType, ValidationRule, DialobItem } from "./types";
-
-export type EditorAction =
-  | { type: 'setActivePage', page: DialobItem }
-  | { type: 'setActiveFormLanguage', language: string }
-
+import { DialobItemTemplate, ValueSetEntry, ContextVariableType, ValidationRule } from "./types";
 
 export type ComposerAction =
   | { type: 'addItem', config: DialobItemTemplate, parentItemId: string, afterItemId?: string}
@@ -37,4 +32,3 @@ export type ComposerAction =
 
   | { type: 'addLanguage', language: string, copyFrom?: string}
   | { type: 'deleteLanguage', language: string}
-
