+++ frontend/dialob-composer-material/src/dialob/index.ts
@@ -1,7 +1,3 @@
-// Public export from dialob composer core
-
 export { useComposer } from "./react/useComposer"
-export { useEditor } from "./react/useEditor"
 export { ComposerContext, ComposerProvider} from './react/ComposerContext';
-export { EditorContext, EditorProvider } from './react/EditorContext';
 export * from './types';
