+++ frontend/dialob-composer-material/src/editor/actions.ts
@@ -1,0 +1,5 @@
+import { DialobItem } from "../dialob";
+
+export type EditorAction =
+  | { type: 'setActivePage', page: DialobItem }
+  | { type: 'setActiveFormLanguage', language: string }
