+++ frontend/dialob-composer-material/src/editor/types.ts
@@ -1,0 +1,6 @@
+import { DialobItem } from "../dialob";
+
+export type EditorState = {
+  activePage?: DialobItem;
+  activeFormLanguage: string;
+};
