+++ frontend/dialob-composer-material/src/views/tree/NavigationTreeItem.tsx
@@ -55,7 +55,7 @@ const NavigationTreeItem: React.FC<TreeItemProps> = ({ item, onExpand, onCollaps
     >
       {getIcon(item, onExpand, onCollapse)}
       {getTypeIcon(item.data.item, item.data.isPage)}
-      <ListItemText sx={{ cursor: 'pointer' }}>{item.data ? item.data.title : ''}</ListItemText>
+      <ListItemText sx={{ cursor: 'pointer', '.MuiListItemText-primary:hover': { fontWeight: 'inherit' } }}>{item.data ? item.data.title : ''}</ListItemText>
     </ListItem>
   );
 };
